requirements.txt
# Core Python dependencies for Schwab Trading Agent Tools
# Minimum Python version: 3.11

# HTTP Client
requests>=2.31.0

# Data manipulation and analysis
pandas>=2.0.0
numpy>=1.24.0

# Data validation
pydantic>=2.0.0

# Configuration management
python-dotenv>=1.0.0

# Windows-specific: Token encryption using DPAPI
pywin32>=306

# OAuth 2.0
authlib>=1.2.0

# Logging and utilities
python-dateutil>=2.8.0
